Add a build system (CMake)

+++ src/core/types.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <cstdint>
+#include <cstddef>
 
 using i32 = int32_t;
 using u32 = uint32_t;
@@ -10,5 +11,5 @@ using f32 = float;
 using f64 = double;
 
 using label = i32;
-using ulabel = u32;
+using ulabel = size_t;
 using scalar = f64;
